Use IconButtonWrapper for EntityItem's icon
